Moved the fall detection to background service
